// src/app/model/customer.ts
export class Customer {
    id: string;
    customerName: string;
    countryId : number;
    provinceId : number;
    address : string;
    email : string;
    phoneNumber : string;
    userName : string;
    userPassword : string;
    timeZone : string;
    roleId: number;
    languageId: number;
    timezoneId: number;
    password: string;
    confirmPassword: string;
  }